queue_keep and init_delay now adjustable values in hrcheck.py

+++ hrcheck.py
@@ -22,11 +22,13 @@ import os
 import datetime
 import re
 import calendar
+import time
 
 dupes = ['', 's']
 
 from collections import defaultdict
 
+init_delay = 0
 hour_parts = 4
 
 of_day = defaultdict(str)
@@ -69,7 +71,8 @@ def usage():
     print("=" * 50)
     print("hh = normalizes to half hour e.g. :12 or :18 look for both :00 and :15 tipoffs. nh/hn turns it off.")
     print("rp/p/r decides whether to print or run current commands")
-    print("l/ul/lu = lock the lockfile, u = unlock the lockfile, q = run the queue file, lq = list queue")
+    print("l/ul/lu = lock the lockfile, u = unlock the lockfile, q = run the queue file, lq = list queue, kq/qk=keep queue file")
+    print("id specifies initial delay")
     exit()
 
 def garbage_collect(x):
@@ -169,7 +172,7 @@ def carve_neg(ti):
     return retval
 
 def see_what_to_run(ti, wd, md, hh):
-    # print(ti, "= time index", wd, "= weekday index", md, "=monthday index", hh, "=whether to go in same half hour")
+    print(ti, "= time index", wd, "= weekday index", md, "=monthday index", hh, "=whether to go in same half hour")
     totals = 0
     totals += carve_neg(ti)
     totals += carve_up(of_day[ti], "daily run on")
@@ -254,11 +257,13 @@ def run_queue_file():
                 continue
             see_what_to_run(my_list[0], my_list[1], my_list[2], half_hour)
     f.close()
-    f = open(queue_file, "w")
-    f.write("#queue file format = (time index),(day of week),(day of month)\n")
-    f.close()
+    if not queue_keep:
+        f = open(queue_file, "w")
+        f.write("#queue file format = (time index),(day of week),(day of month)\n")
+        f.close()
 
 queue_run = 0
+queue_keep = True
 count = 1
 half_hour = False
 run_cmd = True
@@ -288,9 +293,15 @@ while count < len(sys.argv):
     elif arg == 'q':
         unlock_lock_file(False)
         queue_run = 1
+    elif arg == 'qk' or arg == 'kq':
+        unlock_lock_file(False)
+        queue_run = 1
+        queue_keep = False
     elif arg == 'ql' or arg == 'lq':
         list_queue()
         exit()
+    elif arg[0] == 'id':
+        init_delay = int(arg[2:])
     elif arg[0] == 'm':
         mday = int(arg[1:])
     elif arg[0] == 'w':
@@ -330,6 +341,8 @@ if file_lock():
 read_hourly_check(check_file)
 read_hourly_check(check_private)
 read_hourly_check(xtra_file)
+
+if init_delay: sleep(init_delay)
 
 if queue_run == 1:
     run_queue_file()
